pyproject.toml
[project]
name = "image-video-parser"
version = "0.1.0"
description = ""
authors = [
    {name = "Domenico Bini"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "torch (>=2.5.1,<3.0.0)",
    "torchvision (>=0.20.1,<0.21.0)",
    "transformers (>=4.48.0,<5.0.0)",
    "llama-index (>=0.12.10,<0.13.0)",
    "llama-index-utils-workflow (>=0.3.0,<0.4.0)",
    "llama-index-multi-modal-llms-azure-openai (>=0.3.1,<0.4.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "sam-2 @ git+ssh://git@github.com/facebookresearch/sam2.git",
    "accelerate (>=1.2.1,<2.0.0)",
    "scipy (>=1.15.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
